Wait for sd card to (re)appear to get short filename.
We need to get short filename for long filename that we know.
Unfortunately refresh_sd_files(blocking=True) doesn't wait for data if
it sees that sd card is not yet available.

We have to wait ourselves. Fortunately we do that in our own thread, so
we won't stop octoprint main thread. We wait for up to 10s.